Moved ColumnStatsMerger and implementors.
